Move quick fix classes to the new quickfix package
