+++ com.reprezen.swagedit.core/META-INF/MANIFEST.MF
@@ -25,6 +25,7 @@ Export-Package: com.reprezen.swagedit.core.assist,
  com.reprezen.swagedit.core.json.references,
  com.reprezen.swagedit.core.model,
  com.reprezen.swagedit.core.preferences,
+ com.reprezen.swagedit.core.quickfix,
  com.reprezen.swagedit.core.schema,
  com.reprezen.swagedit.core.templates,
  com.reprezen.swagedit.core.utils,
