+++ com.reprezen.swagedit/plugin.xml
@@ -199,7 +199,7 @@
    <extension point="org.eclipse.ui.ide.markerResolution">
       <markerResolutionGenerator
          markerType="org.eclipse.core.resources.problemmarker"
-         class="com.reprezen.swagedit.validation.SwaggerQuickFixer"/>
+         class="com.reprezen.swagedit.quickfix.SwaggerQuickFixer"/>
    </extension>
    
 </plugin>
